Move all clients to the base class
